python-script/requirements.txt
pyimagej==1.5.0
scyjava==1.12.1
python_dotenv==1.2.1